@@ -66,10 +66,19 @@ function LibraryItem() {
           url = item.electronicLocator[0].url;
           actionLabel = 'View online';
           actionLabelHelper = `resource for ${recordTitle}`;
+          // Temporary for ReCAP items.
         } else if (accessMessage === 'adv request' && !item.holdingLocation) {
           requestHold = true;
           actionLabel = accessMessage;
-          actionLabelHelper = 'request hold on ${recordTitle}';
+          actionLabelHelper = `request hold on ${recordTitle}`;
+          // Temporary for NYPL ReCAP items.
+          // Making sure that if there is a holding location, that the location code starts with
+          // rc. Ids are in the format of `loc:x` where x is the location code.
+        } else if (item.holdingLocation && item.holdingLocation.length &&
+          item.holdingLocation[0]['@id'].substring(4, 6) === 'rc') {
+          requestHold = true;
+          actionLabel = accessMessage;
+          actionLabelHelper = `request hold on ${recordTitle}`;
         } else if (availability === 'available') {
           url = this.getLocationHoldUrl(locationDetails);
           actionLabel = 'Request for in-library use';
