@@ -4,6 +4,19 @@ import appConfig from '../../../appConfig.js';
 const appEnvironment = process.env.APP_ENV || 'production';
 const apiBase = appConfig.api[appEnvironment];
 
+function requireUser(req, res) {
+  if (!req.tokenResponse || !req.tokenResponse.isTokenValid ||
+    !req.tokenResponse.accessToken || !req.tokenResponse.decodedPatron ||
+    !req.tokenResponse.decodedPatron.sub) {
+    // redirect to login
+    const fullUrl = encodeURIComponent(`${req.protocol}://${req.get('host')}${req.originalUrl}`);
+
+    res.redirect(`${appConfig.loginUrl}?redirect_uri=${fullUrl}`);
+    return false;
+  }
+  return true;
+}
+
 function retrieveItem(q, cb, errorcb) {
   return axios
     .get(`${apiBase}/discovery/resources/${q}`)
@@ -13,6 +26,34 @@ function retrieveItem(q, cb, errorcb) {
 
       errorcb(error);
     }); /* end axios call */
+}
+
+function serverConfirmRequest(req, res, next) {
+  const q = req.params.id || '';
+  const loggedIn = requireUser(req, res);
+
+  if (!loggedIn) return false;
+
+  return retrieveItem(
+    q,
+    (data) => {
+      res.locals.data.Store = {
+        bib: data,
+        searchKeywords: '',
+        error: {},
+      };
+      next();
+    },
+    (error) => {
+      console.log(error);
+      res.locals.data.Store = {
+        bib: {},
+        searchKeywords: '',
+        error,
+      };
+      next();
+    }
+  );
 }
 
 function serverItemSearch(req, res, next) {
@@ -52,19 +93,6 @@ function ajaxItemSearch(req, res) {
 
 function account(req, res, next) {
   next();
-}
-
-function requireUser(req, res) {
-  if (!req.tokenResponse || !req.tokenResponse.isTokenValid ||
-    !req.tokenResponse.accessToken || !req.tokenResponse.decodedPatron ||
-    !req.tokenResponse.decodedPatron.sub) {
-    // redirect to login
-    const fullUrl = encodeURIComponent(`${req.protocol}://${req.get('host')}${req.originalUrl}`);
-
-    res.redirect(`${appConfig.loginUrl}?redirect_uri=${fullUrl}`);
-    return false;
-  }
-  return true;
 }
 
 function newHoldRequest(req, res, next) {
@@ -108,7 +136,7 @@ function createHoldRequest(req, res) {
 
   // get item id and pickup location
   let itemId = req.params.id;
-  let nyplSource = 'nypl-sierra';
+  let nyplSource = 'sierra-nypl';
 
   if (itemId.indexOf('-') >= 0) {
     const parts = itemId.split('-');
@@ -154,10 +182,74 @@ function createHoldRequest(req, res) {
     }); /* end axios call */
 }
 
+function ajaxCreateHoldRequest(req, res) {
+  // Ensure user is logged in
+  const loggedIn = requireUser(req);
+  if (!loggedIn) return false;
+
+  // retrieve access token and patron info
+  const accessToken = req.tokenResponse.accessToken;
+  const patronId = req.tokenResponse.decodedPatron.sub;
+  const patronHoldsApi = `${appConfig.api.development}/hold-requests`;
+
+  // get item id and pickup location
+  let itemId = req.query.id;
+  let nyplSource = 'sierra-nypl';
+
+  if (itemId.indexOf('-') >= 0) {
+    const parts = itemId.split('-');
+    itemId = parts[parts.length - 1];
+
+    if (itemId.substring(0, 2) === 'pi') {
+      nyplSource = 'recap-PUL';
+    } else if (itemId.substring(0, 2) === 'ci') {
+      nyplSource = 'recap-CUL';
+    }
+  }
+  itemId = itemId.replace(/\D/g, '');
+  const pickupLocation = req.body.pickupLocation;
+
+  const data = {
+    patron: patronId,
+    recordType: 'i',
+    record: itemId,
+    nyplSource,
+    pickupLocation,
+    // neededBy: "2013-03-20",
+    numberOfCopies: 1,
+  };
+  // console.log('Making hold request', data, accessToken);
+
+  return axios
+    .post(patronHoldsApi, data, {
+      headers: {
+        'Content-Type': 'application/json',
+        Authorization: `Bearer ${accessToken}`,
+      },
+    })
+    .then(response => {
+      res.json({
+        id: response.data.data.id,
+        jobId: response.data.data.jobId,
+        holdRequest: data,
+      });
+    })
+    .catch(error => {
+      console.log(`Error calling Holds API : ${error.data.message}`);
+      res.json({
+        status: error.status,
+        error,
+      });
+    }); /* end axios call */
+}
+
+
 export default {
   serverItemSearch,
   newHoldRequest,
   createHoldRequest,
   account,
   ajaxItemSearch,
+  ajaxCreateHoldRequest,
+  serverConfirmRequest,
 };
